+++ darwin/Classes/AudioplayersPlugin.m
@@ -55,6 +55,7 @@ NSString *_artist;
 NSString *_imageUrl;
 int _duration;
 const float _defaultPlaybackRate = 1.0;
+const NSString *_defaultPlayingRoute = @"speakers";
 
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
   _registrar = registrar;
@@ -276,6 +277,12 @@ const float _defaultPlaybackRate = 1.0;
                     NSString *releaseMode = call.arguments[@"releaseMode"];
                     bool looping = [releaseMode hasSuffix:@"LOOP"];
                     [self setLooping:looping playerId:playerId];
+                  },
+                @"earpieceOrSpeakersToggle":
+                  ^{
+                    NSLog(@"earpieceOrSpeakersToggle");
+                    NSString *playingRoute = call.arguments[@"playingRoute"];
+                    [self setPlayingRoute:playingRoute playerId:playerId];
                   }
                 };
 
@@ -293,7 +300,7 @@ const float _defaultPlaybackRate = 1.0;
 -(void) initPlayerInfo: (NSString *) playerId {
   NSMutableDictionary * playerInfo = players[playerId];
   if (!playerInfo) {
-    players[playerId] = [@{@"isPlaying": @false, @"volume": @(1.0), @"rate": @(_defaultPlaybackRate), @"looping": @(false)} mutableCopy];
+    players[playerId] = [@{@"isPlaying": @false, @"volume": @(1.0), @"rate": @(_defaultPlaybackRate), @"looping": @(false), @"playingRoute": _defaultPlayingRoute} mutableCopy];
   }
 }
 
@@ -341,6 +348,10 @@ const float _defaultPlaybackRate = 1.0;
           MPRemoteCommand *togglePlayPauseCommand = [remoteCommandCenter togglePlayPauseCommand];
           [togglePlayPauseCommand setEnabled:YES];
           [togglePlayPauseCommand addTarget:self action:@selector(playOrPauseEvent:)];
+
+          MPRemoteCommand *changePlaybackPositionCommand = [remoteCommandCenter changePlaybackPositionCommand];
+          [changePlaybackPositionCommand setEnabled:YES];
+          [changePlaybackPositionCommand addTarget:self action:@selector(onChangePlaybackPositionCommand:)];
         }
     }
 
@@ -376,6 +387,7 @@ const float _defaultPlaybackRate = 1.0;
         }
         return MPRemoteCommandHandlerStatusSuccess;
     }
+
     -(MPRemoteCommandHandlerStatus) playOrPauseEvent: (MPSkipIntervalCommandEvent *) playOrPauseEvent {
         NSLog(@"playOrPauseEvent");
 
@@ -403,6 +415,13 @@ const float _defaultPlaybackRate = 1.0;
         if (headlessServiceInitialized) {
           [_callbackChannel invokeMethod:@"audio.onNotificationBackgroundPlayerStateChanged" arguments:@{@"playerId": _currentPlayerId, @"updateHandleMonitorKey": @(_updateHandleMonitorKey), @"value": playerState}];
         }
+        return MPRemoteCommandHandlerStatusSuccess;
+    }
+
+    -(MPRemoteCommandHandlerStatus) onChangePlaybackPositionCommand: (MPChangePlaybackPositionCommandEvent *) changePositionEvent {
+        NSLog(@"changePlaybackPosition to %f", changePositionEvent.positionTime);
+        CMTime newTime = CMTimeMakeWithSeconds(changePositionEvent.positionTime, NSEC_PER_SEC);
+        [ self seek:_currentPlayerId time:newTime ];
         return MPRemoteCommandHandlerStatusSuccess;
     }
 
@@ -469,8 +488,14 @@ const float _defaultPlaybackRate = 1.0;
         success = [[AVAudioSession sharedInstance] setCategory:category withOptions:AVAudioSessionCategoryOptionMixWithOthers error:&error];
       } else {
         success = [[AVAudioSession sharedInstance] setCategory:category error:&error];
+        [[UIApplication sharedApplication] beginReceivingRemoteControlEvents];
       }
-    
+      
+      if ([playerInfo[@"playingRoute"] isEqualToString:@"earpiece"]) {
+        // Use earpiece speaker to play audio.
+        success = [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayAndRecord error:&error];
+      }
+
       if (!success) {
         NSLog(@"Error setting speaker: %@", error);
       }
@@ -487,7 +512,7 @@ const float _defaultPlaybackRate = 1.0;
     }
       
     if (playerInfo[@"url"]) {
-      [[player currentItem] removeObserver:self forKeyPath:@"player.currentItem.status" ];
+      [[player currentItem] removeObserver:self forKeyPath:@"status" ];
 
       [ playerInfo setObject:url forKey:@"url" ];
 
@@ -523,7 +548,7 @@ const float _defaultPlaybackRate = 1.0;
     // is sound ready
     [playerInfo setObject:onReady forKey:@"onReady"];
     [playerItem addObserver:self
-                          forKeyPath:@"player.currentItem.status"
+                          forKeyPath:@"status"
                           options:0
                           context:(void*)playerId];
       
@@ -668,6 +693,25 @@ recordingActive: (bool) recordingActive
   [playerInfo setObject:@(looping) forKey:@"looping"];
 }
 
+-(void) setPlayingRoute: (NSString *) playingRoute
+               playerId: (NSString *) playerId {
+  NSLog(@"%@ -> calling setPlayingRoute", osName);
+  NSMutableDictionary *playerInfo = players[playerId];
+  [playerInfo setObject:(playingRoute) forKey:@"playingRoute"];
+
+  BOOL success = false;
+  NSError *error = nil;
+  if ([playingRoute isEqualToString:@"earpiece"]) {
+    // Use earpiece speaker to play audio.
+    success = [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayAndRecord error:&error];
+  } else {
+    success = [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback error:&error];
+  }
+  if (!success) {
+    NSLog(@"Error setting playing route: %@", error);
+  }
+} 
+
 -(void) stop: (NSString *) playerId {
   NSMutableDictionary * playerInfo = players[playerId];
 
@@ -734,7 +778,7 @@ recordingActive: (bool) recordingActive
                      ofObject:(id)object
                        change:(NSDictionary *)change
                       context:(void *)context {
-  if ([keyPath isEqualToString: @"player.currentItem.status"]) {
+  if ([keyPath isEqualToString: @"status"]) {
     NSString *playerId = (__bridge NSString*)context;
     NSMutableDictionary * playerInfo = players[playerId];
     AVPlayer *player = playerInfo[@"player"];