+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
 ## [next]
+- Implemented stream routing for iOS
+- Call release on dispose
+- Fix iOS build
+- Breaking change audio cache prefix in order to allow override 'assets'
+
+## audioplayers 0.15.1
+- Fix web for release mode
+
+## audioplayers 0.15.0
+- Improve loop/readme for web support
+- Audio cache support for web
+- Re-adding partial web support
 
 ## audioplayers 0.14.2
 - Fix pubspec problem because of web file